Sign up page + code refactoring for home
