+++ src/Pages/Signup/signup.css
@@ -1,0 +1,91 @@
+.auth-container {
+    display: flex;
+    width: 200vh;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+    background-color: #f5f5f5;
+}
+
+.auth-form {
+    background: #fff;
+    padding: 40px;
+    border-radius: 8px;
+    box-shadow: 0 2px 10px #dad6d61a;
+    text-align: center;
+    width: 350px;
+}
+
+.auth-form h1 {
+    font-family:'Lucida Sans', 'Lucida Sans Regular', 'Lucida Grande', 'Lucida Sans Unicode', Geneva, Verdana, sans-serif;
+    margin-bottom: 10px;    
+    font-size: 30px;
+    font-weight: bold;
+    color: black;
+}
+
+.auth-form p {
+    margin: 5px 0;
+    color: #666;
+}
+
+.auth-form form {
+    display: flex;
+    flex-direction: column;
+}
+
+.auth-form input {
+    padding: 10px;
+    margin: 10px 0;
+    border: 1px solid #ccc;
+    border-radius: 2.5px;
+}
+
+.auth-form button {
+    background-color: #4CAF50;
+    color: white;
+    padding: 10px;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    margin-top: 10px;
+}
+
+.auth-form button:hover {
+    background-color: #45a049;
+}
+
+.auth-form .social-login {
+    display: flex;
+    justify-content: space-around;
+    margin: 20px 0;
+}
+
+.auth-form .social-login button {
+    background: none;
+    border: none;
+    cursor: pointer;
+}
+
+.auth-form .google {
+    color: #db4437;
+}
+
+.auth-form .facebook {
+    color: #3b5998;
+}
+
+.auth-form .apple {
+    color: #000;
+}
+
+.auth-form .skip {
+    display: block;
+    margin-top: 20px;
+    color: #1e90ff;
+    text-decoration: none;
+}
+
+.auth-form .skip:hover {
+    text-decoration: underline;
+}
